+++ src/components/index.js
@@ -2,3 +2,5 @@ export { RouterComponent } from './router/router';
 export { VenueMap } from './venueMap/venueMap';
 export { Venue } from './venue/venue';
 export { Header } from './header/header';
+export { ListView } from './listView/listView';
+export { ListItem } from './listItem/listItem';