+++ src/components/venue/venue.js
@@ -1,9 +1,10 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import MdCall from 'react-ionicons/lib/MdCall';
-import Marker from '../../assets/imgs/marker.svg';
-import Type from '../../assets/imgs/type.svg';
+// import MdCall from 'react-ionicons/lib/MdCall';
+// import Marker from '../../assets/imgs/marker.svg';
+// import Type from '../../assets/imgs/type.svg';
 import Default from '../../assets/imgs/default.jpg';
+import { ListView } from '../index';
 
 export class Venue extends React.Component {
 
@@ -24,7 +25,6 @@ export class Venue extends React.Component {
          */
 
         return (
-           
             <div className="card custom-card mt-4">
                 <div className="card-img-holder">
                     <img src={imgSrc ? imgSrc : Default} className="card-img-top" alt="Venue" /> 
@@ -33,24 +33,14 @@ export class Venue extends React.Component {
                     <h5 className="card-title uppercase fw-400 ls-title-">{name}</h5>
                    
                     <p>{description}</p>
-        
-                    <ul className="list-group custom-list-group list-group-flush mb-3">
-                        <li className="list-group-item">
-                            <img src={Marker} alt="address" className="svg-icon-left"/>
-                            { address }
-                        </li>
-                        <li className="list-group-item">
-                            <img src={Type} alt="type" className="svg-icon-left"/>
-                            {categoryName }
-                        </li>
-                        <li className="list-group-item">
-                            <MdCall fontSize="20px" color="#ffffff"  style={{marginRight: '10px', marginTop: '-5px'}}/>
-                            {phone}
-                        </li>
-                    </ul>
+
+                    <ListView 
+                        address={address}
+                        categoryName={categoryName}
+                        phone={phone}
+                    />
                 </div>
-            </div>
-              
+            </div>  
         );
     }
 }